backend/pyproject.toml
[project]
name = "moss-stack-backend"
version = "0.1.0"
description = "Backend for the Legal Proofreader application."
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlmodel",
    "asyncpg",
    "greenlet",
    "alembic",
    "celery",
    "redis",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "structlog>=25.4.0",
    "bcrypt>=4.3.0",
    "aiofiles>=24.1.0",
    "dependency-injector>=4.48.2",
    "confluent-kafka>=2.11.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "factory-boy",
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "mypy",
]
